fix: keep batch transcribe, clipboard snapshots, and data dirs intact
Stop aborting CLI --dir on the first failure, persist pasteboard for crash restore, lock data directories to 0700, and drop the Apple Speech fallback claim.

+++ src-tauri/src/disk.rs
@@ -76,7 +76,7 @@ fn unix_free(path: &Path) -> Option<u64> {
         if libc::statfs(cstr.as_ptr(), &mut buf) != 0 {
             return None;
         }
-        Some((buf.f_bavail as u64).saturating_mul(buf.f_bsize as u64))
+        Some(buf.f_bavail.saturating_mul(buf.f_bsize as u64))
     }
 }
